# Conflicts:
#	Listas/src/Lista01_Revisao/Exercicio19.java
#	Listas/src/Lista01_Revisao/Exercicio20.java
#	out/production/Programação 2/Lista01_Revisao/Exercicio18.class
#	out/production/Programação 2/Lista01_Revisao/Exercicio19.class
#	out/production/Programação 2/Lista01_Revisao/Exercicio20.class

+++ Slides/src/Mes01/Exercicio02_Treino.java
@@ -11,7 +11,7 @@ public class Exercicio02_Treino {
 
         String resultado = "(a + b)^" + n + " = ";
 
-        long coeficiente = 1;
+        int coeficiente = 1;
 
         for (int i = 0; i <= n; i++) {
             if (i > 0) {
